+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 0.2.0
 
 - Adding metering middleware
+- Adding memory pointer to define the allocated memory space
 
 
 ## 0.1.0